API basique sans implémentation
