--- src/kdTreePackage/KdTree.java
@@ -1,24 +1,0 @@
-package kdTreePackage;
-
-public class KdTree {
-	
-	private KdNode racine;
-	// TODO implement racine
-	
-	
-	public KdTree() {
-		
-	}
-	
-	public void addNode( KdNode node ) {
-		
-	}
-	
-	public void removeNode ( KdNode node ) {
-		
-	}
-	
-	public KdNode getNearestNeighbors( KdTree ) {
-		
-	}
-}
